requirement.txt
requests
pytest
allure-pytest
pyyaml
csv